requirements.txt
flask>=2.0.0
flask-sock>=0.5.0
gunicorn>=20.1.0
gevent>=21.8.0
gevent-websocket>=0.10.1
paramiko>=2.8.0
pyyaml>=5.4.1
python-dotenv>=0.19.0
requests>=2.26.0
docker>=6.1.2
watchdog>=2.1.9
typing-extensions>=4.0.0
dataclasses>=0.8; python_version < '3.7'
werkzeug>=2.0.0
click>=8.0.0
itsdangerous>=2.0.0
jinja2>=3.0.0
markupsafe>=2.0.0
six>=1.16.0
urllib3>=1.26.0
certifi>=2021.10.8
charset-normalizer>=2.0.0
idna>=3.3
bcrypt>=3.2.0
cryptography>=3.4.0
pynacl>=1.4.0
cffi>=1.15.0
pycparser>=2.21
zope.event>=4.5.0
zope.interface>=5.4.0
greenlet>=1.1.0
setproctitle>=1.3.2
